requirements.txt
gurobi==10.0.3
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
openai==1.3.0
google-generativeai==0.3.2 